(#58) Fix for Windows

+++ src/shell.cpp
@@ -233,7 +233,7 @@ void Shell::configure_commands_global() {
         std::cout << "List of environments in global:\n";
         for (size_t i = 0; i < envs.size(); ++i) {
             std::cout << "|-> " << envs[i].get_name() << "\n";
-            for (size_t j = 0; j < std::min(3ul, envs[i].get_tasks().size()); ++j) {
+            for (size_t j = 0; j < std::min(size_t(3), envs[i].get_tasks().size()); ++j) {
                 std::cout << "    |-> " << envs[i].get_tasks()[j].get_name() << ": " <<
                     envs[i].get_tasks()[j].get_settings()["language"] << "\n";
             }
